backend/requirements.txt
fastapi
uvicorn
python-multipart
librosa
soundfile
numpy
scipy
Pillow
# moviepy  # Uncomment for video processing support
